requirements.txt
uvicorn==0.16.0
tensorflow-cpu
scikit-learn==1.0.2
numpy==1.22.0
pandas==1.3.5
matplotlib==3.5.1
fastapi==0.71.0
gunicorn
